# refactor: ensure tableName is statically evaluated