Seeing a pretty significant perf increase from this change!

*Before:*

```bash
$ wrk -c 64 -d 30 http://localhost:4000/posts/1

Running 30s test @ http://localhost:4000/posts/1
  2 threads and 64 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    50.53ms   23.54ms 234.52ms   72.10%
    Req/Sec   646.46     82.86   808.00     78.83%
  38676 requests in 30.06s, 44.81MB read
Requests/sec:   1286.49
Transfer/sec:      1.49MB
```

*After:*

```bash
$ wrk -c 64 -d 30 http://localhost:4000/posts/1

Running 30s test @ http://localhost:4000/posts/1
  2 threads and 64 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    43.10ms   19.98ms 191.58ms   72.16%
    Req/Sec   758.54     83.87     1.00k    69.50%
  45367 requests in 30.05s, 52.57MB read
Requests/sec:   1509.88
Transfer/sec:      1.75MB
```